@@ -15,8 +15,8 @@
     },
     "dependencies": {
         "axios": "^0.18.0",
-        "react": "^16.2.0",
-        "react-dom": "^16.2.0",
+        "react": "^16.3.1",
+        "react-dom": "^16.3.1",
         "react-redux": "^5.0.7",
         "react-router-dom": "^4.2.2",
         "react-scripts": "1.1.1",
